Feat | Add max worker count when extracting manifests
