+++ pkg/duplicates/utils.go
@@ -1,4 +1,4 @@
-package argoapplication
+package duplicates
 
 import (
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
